Move OpenACC SharedAllocationRecord implementation to separate header and source file
