+++ core/src/OpenACC/Kokkos_OpenACCSpace.hpp
@@ -52,9 +52,7 @@ static_assert(false,
 #define KOKKOS_OPENACC_SPACE_HPP
 
 #include <Kokkos_Concepts.hpp>
-
 #include <impl/Kokkos_Tools.hpp>
-#include <impl/Kokkos_SharedAlloc.hpp>
 
 #include <openacc.h>
 #include <iosfwd>
@@ -136,65 +134,5 @@ struct Kokkos::Impl::MemorySpaceAccess<Kokkos::Experimental::OpenACCSpace,
   enum : bool { deepcopy = true };
 };
 /*--------------------------------------------------------------------------*/
-
-template <>
-class Kokkos::Impl::SharedAllocationRecord<Kokkos::Experimental::OpenACCSpace,
-                                           void>
-    : public HostInaccessibleSharedAllocationRecordCommon<
-          Kokkos::Experimental::OpenACCSpace> {
- private:
-  friend class HostInaccessibleSharedAllocationRecordCommon<
-      Kokkos::Experimental::OpenACCSpace>;
-  friend class SharedAllocationRecordCommon<Kokkos::Experimental::OpenACCSpace>;
-  friend Kokkos::Experimental::OpenACCSpace;
-
-  using base_t = HostInaccessibleSharedAllocationRecordCommon<
-      Kokkos::Experimental::OpenACCSpace>;
-  using RecordBase = SharedAllocationRecord<void, void>;
-
-  SharedAllocationRecord(const SharedAllocationRecord&) = delete;
-  SharedAllocationRecord& operator=(const SharedAllocationRecord&) = delete;
-
-  /**\brief  Root record for tracked allocations from this OpenACCSpace
-   * instance */
-  static RecordBase s_root_record;
-
-  const Kokkos::Experimental::OpenACCSpace m_space;
-
- protected:
-  ~SharedAllocationRecord();
-  SharedAllocationRecord() = default;
-
-  template <typename ExecutionSpace>
-  SharedAllocationRecord(
-      const ExecutionSpace& /*exec_space*/,
-      const Kokkos::Experimental::OpenACCSpace& arg_space,
-      const std::string& arg_label, const size_t arg_alloc_size,
-      const RecordBase::function_type arg_dealloc = &deallocate)
-      : SharedAllocationRecord(arg_space, arg_label, arg_alloc_size,
-                               arg_dealloc) {}
-
-  SharedAllocationRecord(
-      const Kokkos::Experimental::OpenACC& exec_space,
-      const Kokkos::Experimental::OpenACCSpace& arg_space,
-      const std::string& arg_label, const size_t arg_alloc_size,
-      const RecordBase::function_type arg_dealloc = &deallocate);
-
-  SharedAllocationRecord(
-      const Kokkos::Experimental::OpenACCSpace& arg_space,
-      const std::string& arg_label, const size_t arg_alloc_size,
-      const RecordBase::function_type arg_dealloc = &deallocate);
-
- public:
-  KOKKOS_INLINE_FUNCTION static SharedAllocationRecord* allocate(
-      const Kokkos::Experimental::OpenACCSpace& arg_space,
-      const std::string& arg_label, const size_t arg_alloc_size) {
-    if (acc_on_device(acc_device_host)) {
-      return new SharedAllocationRecord(arg_space, arg_label, arg_alloc_size);
-    } else {
-      return nullptr;
-    }
-  }
-};
 
 #endif
